Create a module to bootstrap fluxcd
A wrapper module with opinionated defaults to provision fluxcd to EKS
cluster with separate credentials for flux controller, so flux can be
limited to read only access.

FluxCD ServiceAccount is annotated with given role that's expected to be
IRSA role with access to KMS that's used to decrypt SOPS encrypted
secrets.

Additionally there's a convenience support to supply key-value pairs to
create a ConfigMap, which then can be used for flux postBuild. Serves as
a way to pass information from terraform to fluxcd.