Encode labels and annotations in json for Kustomization
Encoding them with yaml will mean that the annotations variable will end
up being something like this:

```yaml
foo: bar
bar: baz
```

When injected into the Kustomization however, that's not sufficient,
because it will result in the following:

```yaml
patches:
- patch: |
    apiVersion: v1
    kind: ServiceAccount
    metadata:
      name: kustomize-controller
      annotations: "bar": "baz"
"foo": "bar"
```

Which is of course not valid. JSON encoding solves that because YAML is
a superset of JSON, so any JSON is valid YAML, and it doesn't rely on
indentation but can just be used inline, resulting in the following:

```yaml
patches:
- patch: |
    apiVersion: v1
    kind: ServiceAccount
    metadata:
      name: kustomize-controller
      annotations: {"bar": "baz", "foo": "bar"}
```

Which is actually valid.