Remove the lifecycle rule, manage fluxcd with terraform exclusively
Seems the ignoring lifecycle rules do not work. The original thought was,
`flux_bootstrap_git` would have been used to only FluxCD bootstrapping, as also
the name might suggest. But it turns out if we ignore changes and change
something in the `kustomization_override` on terraform side, the state does not
get updated even with `-refresh-only` and as a result we'd be re-applying the
initial values - for example IRSA role arn or Azure workload identity etc.

So lets remove the lifecycle rule completely and accept we have to manage flux
with terraform, as recommended in fluxcd/terraform-provider-flux#477 - as
`flux_bootstrap_git` is used.